+++ src/Warp.NET/Content/Binaries/Data/TextureData.cs
@@ -1,3 +1,0 @@
-namespace Warp.NET.Content.Binaries.Data;
-
-public record TextureData(ushort Width, ushort Height, byte[] ColorData);