+++ src/Warp.NET/Content/Conversion/Binaries/Data/MeshBinaryData.cs
@@ -1,0 +1,4 @@
+namespace Warp.NET.Content.Conversion.Binaries.Data;
+
+// TODO: Do not re-use Face struct.
+public record MeshBinaryData(string MaterialName, IReadOnlyList<Face> Faces);