+++ src/Warp.NET/Content/Conversion/ContentType.cs
@@ -1,0 +1,9 @@
+namespace Warp.NET.Content.Conversion;
+
+public enum ContentType : byte
+{
+	Shader = 0x00,
+	Texture = 0x01,
+	Sound = 0x02,
+	Model = 0x03,
+}