+++ src/Warp.NET/Content/Conversion/Data/MeshData.cs
@@ -1,0 +1,4 @@
+namespace Warp.NET.Content.Conversion.Data;
+
+// TODO: Do not re-use Face struct.
+public record MeshData(string MaterialName, IReadOnlyList<Face> Faces);