+++ src/Warp.NET/Content/Conversion/DecompiledContentFile.cs
@@ -1,0 +1,3 @@
+namespace Warp.NET.Content.Conversion;
+
+public record DecompiledContentFile(IReadOnlyDictionary<string, Model> Models, IReadOnlyDictionary<string, Shader> Shaders, IReadOnlyDictionary<string, Sound> Sounds, IReadOnlyDictionary<string, Texture> Textures);