+++ src/Warp.NET/_Imports.cs
@@ -5,7 +5,6 @@ global using System.Diagnostics;
 global using System.IO;
 global using System.Linq;
 global using System.Numerics;
-global using System.Reflection;
 global using System.Runtime.CompilerServices;
 global using System.Text;
 global using System.Threading;