+++ CHANGELOG.md
@@ -2,13 +2,18 @@
 
 This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## TODO
+## 0.1.1-alpha.1
 
 ### Fixed
 
 - Fixed incorrect namespaces in source-generated code.
 
 ### Changed
+
+- No longer using reflection to assign content.
+- `Bootstrapper.CreateGame` now requires the content container types as type parameters.
+
+### Removed
 
 - Removed NoahStolk.WaveParser dependency.